Grandes Modificações na interface e melhoramento do código
