+++ 2020-Janeiro-Leitura-Organica/rickecr/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Header from "./Components/Header";
-import Main from "./Components/Main";
+import Header from "./components/Header";
+import Main from "./components/Main";
 
 import './app.css';
