--- 2020-Janeiro-Leitura-Organica/rickecr/src/Components/Main/index.css
@@ -1,44 +1,0 @@
-.title {
-    text-align: center;
-    color: #9EDFA1;
-    font-size: 48px;
-    margin-top: 53px;
-}
-
-.text {
-    display: flex;
-    color: #FFFFFF;
-    font-size: 24px;
-    margin-left: 53px;
-    margin-right: 53px;
-    margin-top: 68px;
-}
-
-.div-btn {
-    display: flex;
-    justify-content: center;
-    margin-top: 56px;
-}
-
-.btn-result {
-    border: 0;
-    margin-top: 30px;
-    background: #9EDFA1;
-    border-radius: 8px;
-    padding: 15px 20px;
-    font-size: 16px;
-    font-weight: bold;
-    color: #000000;
-    cursor: pointer;
-    transition: background 0.5s;
-}
-
-.div-text {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.text {
-    color: #FFFFFF;
-}
