+++ 2020-Janeiro-Leitura-Organica/rickecr/src/components/Header/index.css
@@ -3,7 +3,12 @@
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
-    background-color: #9EDFA1;
+    background-color: #3ab7f1;
+    box-shadow: 0px 2px 4px -1px rgba(0, 0, 0, 0.2);
+    position: fixed;
+    top: 0;
+    z-index: 1;
+    width: 100%;
     height: 75px;
 }
 
@@ -16,6 +21,10 @@
     display: flex;
     flex-direction: row;
     margin-right: 25px;
+}
+
+.div-buttons span {
+    color: #FFFFFF;
 }
 
 .btn-result {
