+++ 2020-Janeiro-Leitura-Organica/rickecr/src/components/Header/index.js
@@ -10,14 +10,18 @@ function Header() {
         <div className="div-header">
             <div className="div-logo">
                 <a href="https://www.leituraorganica.com.br/" target="_blank" rel="noopener noreferrer">
-                    <img src={"https://github.com/codigofalado/desafio333/blob/master/2020-Janeiro-Leitura-Organica/assets/logos/Logo%20-%20Azul.png?raw=true"} alt="logo de leitura orgânica" />
+                    <img src={"https://raw.githubusercontent.com/codigofalado/desafio333/master/2020-Janeiro-Leitura-Organica/assets/logos/Logo%20-%20Preta.png"} alt="logo de leitura orgânica" />
                 </a>
             </div>
 
             <div className="div-buttons">
                 <ButtonGroup variant="text" color="primary" aria-label="text primary button group">
-                    <Button href="#div-info">Sobre</Button>
-                    <Button href="#div-ppm">Calcular PPM</Button>
+                    <Button href="#div-info">
+                        <span>Sobre</span>
+                    </Button>
+                    <Button href="#div-ppm">
+                        <span>Calcular PPM</span>
+                    </Button>
                 </ButtonGroup>
             </div>
         </div>
