+++ 2020-Janeiro-Leitura-Organica/rickecr/src/components/ResultDialog/index.js
