+++ 2020-Janeiro-Leitura-Organica/rickecr/src/components/VelLeitura/index.css
@@ -1,0 +1,40 @@
+.div-btn {
+    display: flex;
+    justify-content: center;
+    margin-top: 56px;
+}
+
+.btn {
+    font-family: 'Lato', monospace;
+    background-color: #21d6a0;
+    border-width: 0;
+    border-radius: 5px;
+    width: 200px;
+    height: 50px;
+    max-width: 100%;
+    margin-bottom: 40px;
+    box-shadow: 0px 1px 5px 0px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 3px 1px -2px rgba(0, 0, 0, 0.12)
+}
+
+.btn-result {
+    border: 0;
+    margin-top: 30px;
+    background-color: #21d6a0;
+    border-radius: 8px;
+    padding: 15px 20px;
+    font-size: 16px;
+    font-weight: bold;
+    color: #000000;
+    cursor: pointer;
+    transition: background 0.5s;
+}
+
+.div-text {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+#div-text {
+    padding-top: 50px;
+}
